refactor(app/routes): use remix v2 route convention
This patch updates my route definitions to conform with the updated v2
file-based routing convention.

See: https://remix.run/docs/en/main/pages/v2#upgrading-to-the-new-convention
