+++ remix.config.js
@@ -2,6 +2,9 @@
  * @type {import('@remix-run/dev').AppConfig}
  */
 module.exports = {
+  future: {
+    v2_routeConvention: true,
+  },
   cacheDirectory: './node_modules/.cache/remix',
   ignoredRouteFiles: ['**/.*', '**/*.css', '**/*.test.{js,jsx,ts,tsx}'],
   serverDependenciesToBundle: ['nanoid/non-secure'],
